Fix for a psycopg2 bug.
Does not manage to handle some fractional seconds properly, when converting
from ticks to an internal timestamp object (potentially datetime.datetime). If
we try to convert a tick value that ends up with a fractional second part
in the interval [59.5, 60.0), the conversion *will* fail internally in
psycopg2. The easiest way of dealing with it is to make sure that
TimestampFromTicks() operates on ints. After all, we do not need proper
sub-second resolution.

+++ Cerebrum/modules/bofhd/utils.py
@@ -318,7 +318,7 @@ class BofhdRequests(object):
             now = time.time()
             if t < now:
                 t = now
-            when = self._db.TimestampFromTicks(t + minutes*60)
+            when = self._db.TimestampFromTicks(int(t + minutes*60))
             self._db.execute("""
                 UPDATE [:table schema=cerebrum name=bofhd_request]
                 SET run_at=:when WHERE request_id=:id""",
